// migrations/snapshots/1a6f807363fb8e1592b3c6851b3a699d082ca4699966af3941e84e3d07920802/contract.d.ts
export type StorageHash =
  StorageHashBase<'1a6f807363fb8e1592b3c6851b3a699d082ca4699966af3941e84e3d07920802'>;
export type ExecutionHash = ExecutionHashBase<string>;
export type ProfileHash =
  ProfileHashBase<'3916f444a8a17ad749191acf9e08dad97d1a327b88c2f1d45d12f240296aa8b2'>;

export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgTargetQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Url: {
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly originalUrl: CodecTypes['pg/text@1']['output'];
      readonly shortCode: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Url: {
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly originalUrl: CodecTypes['pg/text@1']['input'];
      readonly shortCode: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly Url: {
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly originalUrl: CodecTypes['pg/text@1']['output'];
      readonly shortCode: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly Url: {
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly originalUrl: CodecTypes['pg/text@1']['input'];
      readonly shortCode: CodecTypes['pg/text@1']['input'];
    };
  };
};

export namespace Models {
  export type public_Url = {
    createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
    id: CodecTypes['pg/int4@1']['output'];
    originalUrl: CodecTypes['pg/text@1']['output'];
    shortCode: CodecTypes['pg/text@1']['output'];
    readonly [RelationKeys]?: never;
  };
}

export declare const models: {
  public: {
    Url: Models.public_Url;
  };
};

export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;

type ContractBase = Omit<
  ContractType<{
    readonly namespaces: {
      readonly public: {
        readonly id: 'public';
        readonly kind: 'postgres-schema';
        readonly entries: {
          readonly table: {
            readonly Url: {
              columns: {
                readonly createdAt: {
                  readonly nativeType: 'timestamptz';
                  readonly codecId: 'pg/timestamptz-string@1';
                  readonly nullable: false;
                  readonly default: { readonly kind: 'function'; readonly expression: 'now()' };
                };
                readonly id: {
                  readonly nativeType: 'int4';
                  readonly codecId: 'pg/int4@1';
                  readonly nullable: false;
                  readonly default: {
                    readonly kind: 'function';
                    readonly expression: 'autoincrement()';
                  };
                };
                readonly originalUrl: {
                  readonly nativeType: 'text';
                  readonly codecId: 'pg/text@1';
                  readonly nullable: false;
                };
                readonly shortCode: {
                  readonly nativeType: 'text';
                  readonly codecId: 'pg/text@1';
                  readonly nullable: false;
                };
              };
              primaryKey: { readonly columns: readonly ['id'] };
              uniques: readonly [{ readonly columns: readonly ['shortCode'] }];
              indexes: readonly [];
              foreignKeys: readonly [];
            };
          };
        };
      };
    };
    readonly storageHash: StorageHash;
  }>,
  'roots' | 'domain'
> & {
  readonly target: 'postgres';
  readonly targetFamily: 'sql';
  readonly roots: {
    readonly Url: { readonly namespace: 'public' & NamespaceId; readonly model: 'Url' };
  };
  readonly domain: {
    readonly namespaces: {
      readonly public: {
        readonly models: {
          readonly Url: {
            readonly fields: {
              readonly createdAt: {
                readonly nullable: false;
                readonly type: {
                  readonly kind: 'scalar';
                  readonly codecId: 'pg/timestamptz-string@1';
                };
              };
              readonly id: {
                readonly nullable: false;
                readonly type: { readonly kind: 'scalar'; readonly codecId: 'pg/int4@1' };
              };
              readonly originalUrl: {
                readonly nullable: false;
                readonly type: { readonly kind: 'scalar'; readonly codecId: 'pg/text@1' };
              };
              readonly shortCode: {
                readonly nullable: false;
                readonly type: { readonly kind: 'scalar'; readonly codecId: 'pg/text@1' };
              };
            };
            readonly relations: Record<string, never>;
            readonly storage: {
              readonly table: 'Url';
              readonly namespaceId: 'public';
              readonly fields: {
                readonly createdAt: { readonly column: 'createdAt' };
                readonly id: { readonly column: 'id' };
                readonly originalUrl: { readonly column: 'originalUrl' };
                readonly shortCode: { readonly column: 'shortCode' };
              };
            };
          };
        };
      };
    };
  };
  readonly capabilities: {
    readonly postgres: {
      readonly distinctOn: true;
      readonly jsonAgg: true;
      readonly lateral: true;
      readonly limit: true;
      readonly orderBy: true;
      readonly returning: true;
    };
    readonly sql: {
      readonly checkConstraint: true;
      readonly defaultInInsert: true;
      readonly enums: true;
      readonly insertOnConflictSkip: true;
      readonly insertOnConflictWithoutTarget: true;
      readonly lateral: true;
      readonly returning: true;
      readonly scalarList: true;
    };
  };
  readonly extensions: {};
  readonly meta: {};

  readonly profileHash: ProfileHash;
};

export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;